Refactor two-dimensional array printing
compareSorting() now takes in two arrays, calls mergesort/quicksort on both arrays and prints the counters for both with "Counter for mergesort" or "Counter for quicksort" as a prefix for their respective cases.

runComparisons() now calls each sort comparison method 5 times with a header text detailing which ordering of elements is currently being run through and which run the counters are from.

Still need to fix  subsequence generation helper functions not working properly.